refactor server module layering
