+++ server/bootstrap/cli.rs
