+++ server/control/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod control_client;
+pub(crate) mod hot_reload;
+pub(crate) mod local_auth;
+pub(crate) mod null_stores;
+pub(crate) mod service;
