+++ server/ingress/plugins/tcp_pass/mod.rs
@@ -6,20 +6,19 @@ use tracing::{debug, warn};
 use tunnel_lib::plugin::{IngressProtocolHandler, ProtocolKind, Route, ServerCtx};
 use tunnel_lib::ProxyError;
 
-use crate::registry::SharedRegistry;
+use crate::ingress::registry::SharedRegistry;
 
-/// Handles HTTP/1.x and WebSocket connections using byte-level forwarding.
+/// Passthrough TCP handler: forwards raw bytes without protocol inspection.
 ///
-/// Replays the peeked preface bytes into the QUIC tunnel so the client sees
-/// the full original request.
-pub struct H1Handler {
+/// Used for opaque TLS or any unrecognised protocol (ProtocolKind::Tcp).
+pub struct TcpPassHandler {
     pub registry: SharedRegistry,
 }
 
 #[async_trait]
-impl IngressProtocolHandler for H1Handler {
+impl IngressProtocolHandler for TcpPassHandler {
     fn protocol_kind(&self) -> ProtocolKind {
-        ProtocolKind::Http1
+        ProtocolKind::Tcp
     }
 
     async fn handle(
@@ -29,26 +28,25 @@ impl IngressProtocolHandler for H1Handler {
         ctx: &ServerCtx,
     ) -> Result<()> {
         let route = route.ok_or_else(ProxyError::routing_missing_info)?;
-        let hint = ctx
-            .hint
-            .as_ref()
-            .ok_or_else(ProxyError::routing_missing_info)?;
+        let hint = ctx.hint.as_ref();
+        let host = hint.and_then(|h| h.sni.clone().or_else(|| h.authority.clone()));
+        let initial_len = hint.map(|h| h.raw_preface.len()).unwrap_or(0);
 
-        let host = hint
-            .authority
-            .clone()
-            .ok_or_else(ProxyError::routing_missing_host)?;
-
-        let protocol = match hint.protocol {
-            tunnel_lib::proxy::core::Protocol::WebSocket => {
-                tunnel_lib::proxy::core::Protocol::WebSocket
-            }
-            _ => tunnel_lib::proxy::core::Protocol::H1,
-        };
-
-        debug!(host = %host, protocol = ?protocol, "plaintext H1/WS, byte-level forwarding");
+        debug!(
+            host = ?host,
+            initial_len = initial_len,
+            "TCP passthrough"
+        );
 
         let (group_id, proxy_name) = (route.group_id, route.proxy_name);
+
+        let routing_info = tunnel_lib::RoutingInfo {
+            proxy_name: proxy_name.to_string(),
+            src_addr: ctx.peer_addr.ip().to_string(),
+            src_port: ctx.peer_addr.port(),
+            protocol: tunnel_lib::proxy::core::Protocol::Tcp,
+            host,
+        };
 
         let mut attempts = 0;
         let max_attempts = 3;
@@ -79,7 +77,7 @@ impl IngressProtocolHandler for H1Handler {
                         group_id = %group_id,
                         attempt = attempts,
                         error = %e,
-                        "failed to open QUIC stream on selected H1 connection, unregistering and retrying"
+                        "failed to open QUIC stream on selected passthrough TCP connection, unregistering and retrying"
                     );
                     self.registry.unregister(&selected.conn_id);
                     if attempts >= max_attempts {
@@ -92,13 +90,6 @@ impl IngressProtocolHandler for H1Handler {
         let mut send = opened.send;
         let recv = opened.recv;
         let _inflight_guard = opened.inflight;
-        let routing_info = tunnel_lib::RoutingInfo {
-            proxy_name: proxy_name.to_string(),
-            src_addr: ctx.peer_addr.ip().to_string(),
-            src_port: ctx.peer_addr.port(),
-            protocol,
-            host: Some(host),
-        };
         tunnel_lib::send_routing_info(&mut send, &routing_info).await?;
         tunnel_lib::proxy::forward_prefixed_to_client(send, recv, stream, ctx.relay_buf_size).await
     }
