+++ server/ingress/registry.rs
